Updated the code and added documentation
